remove support for "down to" in AudeScript's for loop
This makes the for loop easier to use and more useful. The "to" construct is
more general: it now also works for negative step.

The generated code is optimized whenever possible by evaluating the sign of
step variable. When the sign step is not known at parsing time, an extra check
is done at each iteration.