Mark gui-question as experimental in the UI
Gui-question is a good idea and having it in Aude is a good thing.
However, there are a number of things that are wrong with gui-question.
First, it is not thoroughly tested, nay at all, nor maintained.
Some things were broken when the code landed in the master branch.
The code should be simplified and improved.
The execution flow is not easy to understand. Classes and methods are
a bit weird.
The English needs to be proofread.
The code uses document.getElementById on elements that are generated just
before, like this:

        div.appendChild(libD.jso2dom([
            ["div#question-wording", _("Question:")],

            // To put the automaton for the question
            ["div#question-automata-designer"],

            ["div.button-container", [
                ["button#question-validate", _("Validate")],
                ["button#question-display-response", _("Display response")]
            ]],

            // Answer of the user
            ["div#question-answers"],
        ]));

        // Validate the response
        var butValidate = document.getElementById("question-validate");
        butValidate.addEventListener("click", function (e) {
            question.userResponse = getUserResponse(question);
            showAnswer(question);
        });

libD.jso2dom makes it easy to avoid that:

        div.appendChild(libD.jso2dom([
            ["div#question-wording", _("Question:")],

            // To put the automaton for the question
            ["div#question-automata-designer"],

            ["div.button-container", [
                ["button", {"#": "questionValidateButton"} _("Validate")],
                ["button#question-display-response", _("Display response")]
            ]],

            // Answer of the user
            ["div#question-answers"],
        ], refs));

        // Validate the response
        questionValidateButton.addEventListener("click", function (e) {
            question.userResponse = getUserResponse(question);
            showAnswer(question);
        });

Some styles are hard-coded. These styles need to be moved to the main CSS.
For notifications, there is already a way to make them red and green by using the
right kind of notification.

The UI itself has small problems:
 - Automata answers are transparent
 - When a question has been successfully answered, a new question should be offered.

These are some things that need to be done before removing the experimental flag.